refactor: clean __init__.py files
